Added a getTextOutline function.
This function works like fillText(), except that instead of
filling text, it returns the path data for the text.

The path data consists of an array of numbers. The MOVE_TO command is 0
followed by the x,y coordinate. The LINE_TO command is 1 followed by x,y. The
CURVE_TO command is 3 followed by x,y for the first control point, x,y for the
second control point, and x,y for the destination coordinate. Finally, the
CLOSE_PATH command is 3 followed by no other points.